requirements.txt
Django==3.2.7
djangorestframework==3.12.4
Pillow==8.3.2
psycopg2-binary==2.9.1
python-decouple==3.4
pre-commit==2.15.0
drf-yasg==1.20.0
django-filter~=21.1
django-import-export==2.6.1
